+++ packages/vseed/src/pipeline/advanced/chart/pipes/measures/buildMeasuresForDualAxis.ts
@@ -1,5 +1,6 @@
 import type { AdvancedPipe, DualAxis, DualMeasures, MeasureGroup, Measures, MeasureTree } from 'src/types'
 import { isMeasureTreeWithParentId, isMeasureTreeWithChildren } from './utils'
+import { clone } from 'remeda'
 
 export const buildMeasuresForDualAxis: AdvancedPipe = (advancedVSeed, context) => {
   const { vseed } = context as {
@@ -7,13 +8,12 @@ export const buildMeasuresForDualAxis: AdvancedPipe = (advancedVSeed, context) =
   }
 
   // 带Children的指标树, 不进行任何处理
-  if (isMeasureTreeWithChildren(vseed)) {
-    advancedVSeed.measures = vseed.measures
+  if (isMeasureTreeWithChildren(advancedVSeed.measures)) {
     return advancedVSeed
   }
   // 带parentId的指标树, 转换为带children的指标树
-  if (isMeasureTreeWithParentId(vseed)) {
-    advancedVSeed.measures = generateMeasuresByParentId(vseed.measures as Measures)
+  if (isMeasureTreeWithParentId(advancedVSeed.measures)) {
+    advancedVSeed.measures = generateMeasuresByParentId(advancedVSeed.measures as Measures)
     return advancedVSeed
   }
 
@@ -21,7 +21,7 @@ export const buildMeasuresForDualAxis: AdvancedPipe = (advancedVSeed, context) =
    * 既不是带Children的指标树, 也不是带parentId的指标树, 则自动生成指标
    */
   const dualMeasures = vseed.dualMeasures
-    ? vseed.dualMeasures
+    ? clone(vseed.dualMeasures)
     : basicMeasuresToDualMeasures(advancedVSeed.measures || [])
   advancedVSeed.measures = dualMeasuresToMeasureTree(dualMeasures)