+++ packages/vseed/src/pipeline/advanced/chart/pipes/measures/buildMeasuresForScatter.ts
@@ -1,5 +1,6 @@
 import type { AdvancedPipe, MeasureGroup, Measures, MeasureTree, Scatter, ScatterMeasures } from 'src/types'
 import { isMeasureTreeWithParentId, isMeasureTreeWithChildren } from './utils'
+import { clone } from 'remeda'
 
 export const buildMeasuresForScatter: AdvancedPipe = (advancedVSeed, context) => {
   const { vseed } = context as {
@@ -7,13 +8,12 @@ export const buildMeasuresForScatter: AdvancedPipe = (advancedVSeed, context) =>
   }
 
   // 带Children的指标树, 不进行任何处理
-  if (isMeasureTreeWithChildren(vseed)) {
-    advancedVSeed.measures = vseed.measures
+  if (isMeasureTreeWithChildren(advancedVSeed.measures)) {
     return advancedVSeed
   }
   // 带parentId的指标树, 转换为带children的指标树
-  if (isMeasureTreeWithParentId(vseed)) {
-    advancedVSeed.measures = generateMeasuresByParentId(vseed.measures as Measures)
+  if (isMeasureTreeWithParentId(advancedVSeed.measures)) {
+    advancedVSeed.measures = generateMeasuresByParentId(advancedVSeed.measures as Measures)
     return advancedVSeed
   }
 
@@ -22,7 +22,7 @@ export const buildMeasuresForScatter: AdvancedPipe = (advancedVSeed, context) =>
    */
 
   const scatterMeasures = vseed.scatterMeasures
-    ? vseed.scatterMeasures
+    ? clone(vseed.scatterMeasures)
     : basicMeasuresToScatterMeasures(advancedVSeed.measures || [])
   advancedVSeed.measures = scatterMeasuresToMeasureTree(scatterMeasures)