+++ packages/vseed/src/pipeline/advanced/chart/pipes/measures/deleteTooltipMeasure.ts
@@ -1,0 +1,13 @@
+import { deleteMeasureTreeByCallback } from 'src/pipeline/utils'
+import type { AdvancedPipe, Measure } from 'src/types'
+
+export const deleteTooltipMeasure: AdvancedPipe = (advancedVSeed) => {
+  const deleteBy = (measure: Measure) => measure.encoding === 'tooltip'
+
+  const measureTree = deleteMeasureTreeByCallback(advancedVSeed.measures, deleteBy)
+
+  return {
+    ...advancedVSeed,
+    measures: measureTree,
+  }
+}