+++ packages/vseed/src/pipeline/spec/chart/pipes/tooltip/tooltip.ts
@@ -1,11 +1,11 @@
-import { isEmpty } from 'remeda'
-import { autoFormatter, createFormatter, findMeasureById } from '../../../../utils'
+import { isEmpty, uniqueBy } from 'remeda'
+import { autoFormatter, createFormatter, findAllMeasures, findMeasureById } from '../../../../utils'
 import type { Datum, Dimensions, FoldInfo, Locale, Measures, SpecPipe, Tooltip, UnfoldInfo } from 'src/types'
 import { ORIGINAL_DATA } from 'src/dataReshape'
 
 export const tooltip: SpecPipe = (spec, context) => {
   const result = { ...spec }
-  const { advancedVSeed } = context
+  const { advancedVSeed, vseed } = context
   const { measures, datasetReshapeInfo, chartType, locale, dimensions, encoding } = advancedVSeed
   const baseConfig = advancedVSeed.config[chartType] as { tooltip: Tooltip }
   const { tooltip = { enable: true } } = baseConfig
@@ -21,7 +21,14 @@ export const tooltip: SpecPipe = (spec, context) => {
       title: {
         visible: false,
       },
-      content: createMarkContent(encoding.tooltip || [], dimensions, measures, locale, foldInfo, unfoldInfo),
+      content: createMarkContent(
+        encoding.tooltip || [],
+        dimensions,
+        findAllMeasures(vseed.measures),
+        locale,
+        foldInfo,
+        unfoldInfo,
+      ),
     },
     dimension: {
       title: {
@@ -84,8 +91,14 @@ export const createMarkContent = (
   foldInfo: FoldInfo,
   unfoldInfo: UnfoldInfo,
 ) => {
-  const dims = dimensions.filter((item) => tooltip.includes(item.id))
-  const meas = measures.filter((item) => tooltip.includes(item.id))
+  const dims = uniqueBy(
+    dimensions.filter((item) => tooltip.includes(item.id)),
+    (item) => item.id,
+  )
+  const meas = uniqueBy(
+    measures.filter((item) => tooltip.includes(item.id)),
+    (item) => item.id,
+  )
 
   const dimContent = dims.map((item) => ({
     visible: true,